IR-47: Add a method to get tickets based on sprint ids for JIRA

+++ db/base/models/sprint_member_task.go
@@ -10,10 +10,10 @@ type SprintMemberTask struct {
 	SprintMember     SprintMember
 	SprintMemberID   uint `gorm:"not null"`
 	Task             Task
-	TaskID           uint   `gorm:"not null"`
-	TimeSpentMinutes uint   `gorm:"not null"`
-	PointsEarned     uint   `gorm:"not null"`
-	PointsAssigned   uint   `gorm:"not null"`
-	Rating           int8   `gorm:"default:0; not null"`
-	Comment          string `gorm:"type:text"`
+	TaskID           uint    `gorm:"not null"`
+	TimeSpentMinutes uint    `gorm:"not null"`
+	PointsEarned     float64 `gorm:"not null"`
+	PointsAssigned   float64 `gorm:"not null"`
+	Rating           int8    `gorm:"default:0; not null"`
+	Comment          string  `gorm:"type:text"`
 }
